+++ bitswap/server/internal/decision/engine_test.go
@@ -15,7 +15,6 @@ import (
 
 	"github.com/benbjohnson/clock"
 	wl "github.com/ipfs/boxo/bitswap/client/wantlist"
-	"github.com/ipfs/boxo/bitswap/internal/testutil"
 	message "github.com/ipfs/boxo/bitswap/message"
 	pb "github.com/ipfs/boxo/bitswap/message/pb"
 	blockstore "github.com/ipfs/boxo/blockstore"
@@ -23,6 +22,7 @@ import (
 	"github.com/ipfs/go-cid"
 	ds "github.com/ipfs/go-datastore"
 	dssync "github.com/ipfs/go-datastore/sync"
+	"github.com/ipfs/go-test/random"
 	process "github.com/jbenet/goprocess"
 	peer "github.com/libp2p/go-libp2p/core/peer"
 	libp2ptest "github.com/libp2p/go-libp2p/core/test"
@@ -222,8 +222,8 @@ func TestOutboxClosedWhenEngineClosed(t *testing.T) {
 }
 
 func TestPartnerWantHaveWantBlockNonActive(t *testing.T) {
-	alphabet := "abcdefghijklmnopqrstuvwxyz"
-	vowels := "aeiou"
+	const alphabet = "abcdefghijklmnopqrstuvwxyz"
+	const vowels = "aeiou"
 
 	bs := blockstore.NewBlockstore(dssync.MutexWrap(ds.NewMapDatastore()))
 	for _, letter := range strings.Split(alphabet, "") {
@@ -562,7 +562,7 @@ func TestPartnerWantHaveWantBlockNonActive(t *testing.T) {
 }
 
 func TestPartnerWantHaveWantBlockActive(t *testing.T) {
-	alphabet := "abcdefghijklmnopqrstuvwxyz"
+	const alphabet = "abcdefghijklmnopqrstuvwxyz"
 
 	bs := blockstore.NewBlockstore(dssync.MutexWrap(ds.NewMapDatastore()))
 	for _, letter := range strings.Split(alphabet, "") {
@@ -904,7 +904,7 @@ func TestSendReceivedBlocksToPeersThatWantThem(t *testing.T) {
 	e := newEngineForTesting(ctx, bs, &fakePeerTagger{}, "localhost", 0, WithScoreLedger(NewTestScoreLedger(shortTerm, nil, clock.New())), WithBlockstoreWorkerCount(4))
 	e.StartWorkers(ctx, process.WithTeardown(func() error { return nil }))
 
-	blks := testutil.GenerateBlocksOfSize(4, 8*1024)
+	blks := random.BlocksOfSize(4, 8*1024)
 	msg := message.New(false)
 	msg.AddEntry(blks[0].Cid(), 4, pb.Message_Wantlist_Have, false)
 	msg.AddEntry(blks[1].Cid(), 3, pb.Message_Wantlist_Have, false)
@@ -950,7 +950,7 @@ func TestSendDontHave(t *testing.T) {
 	e := newEngineForTesting(ctx, bs, &fakePeerTagger{}, "localhost", 0, WithScoreLedger(NewTestScoreLedger(shortTerm, nil, clock.New())), WithBlockstoreWorkerCount(4))
 	e.StartWorkers(ctx, process.WithTeardown(func() error { return nil }))
 
-	blks := testutil.GenerateBlocksOfSize(4, 8*1024)
+	blks := random.BlocksOfSize(4, 8*1024)
 	msg := message.New(false)
 	msg.AddEntry(blks[0].Cid(), 4, pb.Message_Wantlist_Have, false)
 	msg.AddEntry(blks[1].Cid(), 3, pb.Message_Wantlist_Have, true)
@@ -1016,7 +1016,7 @@ func TestWantlistForPeer(t *testing.T) {
 	e := newEngineForTesting(ctx, bs, &fakePeerTagger{}, "localhost", 0, WithScoreLedger(NewTestScoreLedger(shortTerm, nil, clock.New())), WithBlockstoreWorkerCount(4))
 	e.StartWorkers(ctx, process.WithTeardown(func() error { return nil }))
 
-	blks := testutil.GenerateBlocksOfSize(4, 8*1024)
+	blks := random.BlocksOfSize(4, 8*1024)
 	msg := message.New(false)
 	msg.AddEntry(blks[0].Cid(), 2, pb.Message_Wantlist_Have, false)
 	msg.AddEntry(blks[1].Cid(), 3, pb.Message_Wantlist_Have, false)
@@ -1455,7 +1455,7 @@ func TestTaggingPeers(t *testing.T) {
 }
 
 func TestTaggingUseful(t *testing.T) {
-	peerSampleIntervalHalf := 10 * time.Millisecond
+	const peerSampleIntervalHalf = 10 * time.Millisecond
 
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()